move location functions to hook
