+++ src/components/App.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { clsx } from "clsx";
 import { ILocation } from "../types";
-import { kInitialDimensions, kVersion, kPluginName, kDefaultOnAttributes, kSimulationTabDimensions, kDataContextName, kParentCollectionName } from "../constants";
-import { initializePlugin, codapInterface, getAttribute, getAllItems } from "@concord-consortium/codap-plugin-api";
+import { kInitialDimensions, kVersion, kPluginName, kDefaultOnAttributes, kSimulationTabDimensions } from "../constants";
+import { initializePlugin, codapInterface } from "@concord-consortium/codap-plugin-api";
 import { LocationTab } from "./location-tab";
 import { SimulationTab } from "./simulation-tab";
 import { Header } from "./header";
@@ -35,38 +35,6 @@ export const App: React.FC = () => {
 
     initialize();
   }, []);
-
-  const extractUniqueLocations = (allItems: any): ILocation[] => {
-    const allCases = allItems.values;
-    const uniqueLocations: ILocation[] = [];
-
-    allCases.forEach((c: any) => {
-      const locationObj: ILocation = {
-        name: c.values.location,
-        latitude: c.values.latitude,
-        longitude: c.values.longitude,
-        coordinatePair: `(${c.values.latitude},${c.values.longitude})`
-      };
-
-      if (!uniqueLocations.some((l) => l.coordinatePair === locationObj.coordinatePair)) {
-        uniqueLocations.push(locationObj);
-      }
-    });
-
-    return uniqueLocations;
-  }
-
-
-  const setUserLocations = async () => {
-    const locationAttr = await getAttribute(kDataContextName, kParentCollectionName, "location");
-    if (locationAttr.success){
-      const allItems = await getAllItems(kDataContextName);
-      if (allItems.success){
-        const uniqeLocations: ILocation[] = extractUniqueLocations(allItems.values);
-        setLocations(uniqeLocations);
-      }
-    }
-  };
 
   const handleTabClick = (tab: "location" | "simulation") => {
     setActiveTab(tab);
@@ -101,7 +69,7 @@ export const App: React.FC = () => {
           setSelectedAttributes={setSelectedAttributes}
           setDataContext={setDataContext}
           locations={locations}
-          setUserLocations={setUserLocations}
+          setLocations={setLocations}
         />
       </div>
       <div className={clsx("tab-content", { active: activeTab === "simulation" })}>
